Make JSX Types not global
