+++ packages/dom-expressions/src/runtime.d.ts
@@ -1,3 +1,4 @@
+import { JSX } from "./jsx";
 export const Aliases: Record<string, string>;
 export const Properties: Set<string>;
 export const ChildProperties: Set<string>;
